refactor(M5): Clean up figure scripts with helper functions and type fixes
- Add _create_panel() helper to figure01.py for better organization
- Fix mypy type errors in both figure scripts
  - Proper matplotlib type imports (Axes not plt.Axes)
  - Explicit type annotations for regions and dicts
  - Fix set_bounds() calls to use separate arguments
  - Cast spikes to float64 for plot_combined_diagnostics
- Both figures now type-safe and generate successfully
- All 86 tests passing

figure01.py: 223 lines (clean orchestration with helper)
figure02.py: 196 lines (<200 target met!)

+++ figures/figure02.py
@@ -173,7 +173,7 @@ def run_demo(params: DecodeParams) -> None:
     plot_combined_diagnostics(
         xs,
         x_true,
-        spikes,
+        spikes.astype(np.float64),
         metrics,
         th,
         params,
